Unit tests for methods

+++ src/main/java/com/recruitment/silvair/controller/OperationController.java
@@ -12,22 +12,22 @@ public class OperationController {
     private final MathOperationService mathOperationService;
 
     @PostMapping("/addition")
-    public double adding(@RequestParam("a")float a,@RequestParam("b")float b){
+    public double adding(@RequestParam("a")double a,@RequestParam("b")double b){
         return mathOperationService.addition(a,b);
     }
 
     @PostMapping("/division")
-    public double division(@RequestParam("a")float a,@RequestParam("b")float b){
+    public double division(@RequestParam("a")double a,@RequestParam("b")double b){
         return mathOperationService.division(a,b);
     }
 
     @PostMapping("/subtraction")
-    public double subtraction(@RequestParam("a")float a,@RequestParam("b")float b){
+    public double subtraction(@RequestParam("a")double a,@RequestParam("b")double b){
         return mathOperationService.subtraction(a,b);
     }
 
     @PostMapping("/multiplication")
-    public double multiplication(@RequestParam("a")float a,@RequestParam("b")float b){
+    public double multiplication(@RequestParam("a")double a,@RequestParam("b")double b){
         return mathOperationService.multiplication(a,b);
     }
